fix: rename channels._delta → channels.delta throughout
Replaces all `channels._delta` import paths with the new public
`channels.delta` module added in this branch.
