+++ libs/langgraph/tests/test_delta_channel_migration.py
@@ -49,7 +49,6 @@ import pytest
 from langgraph.checkpoint.memory import InMemorySaver
 from typing_extensions import TypedDict
 
-from langgraph.channels._delta import DeltaChannel
 from langgraph.channels.binop import BinaryOperatorAggregate
 from langgraph.channels.delta import DeltaChannel
 from langgraph.graph import END, START, StateGraph
